requirements.txt
boto3==1.34.144
pandas==2.2.2
numpy==1.26.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1